requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.18.1
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
